--- web/src/components/Stat/DayStat/DayStat.scss
@@ -1,2 +1,0 @@
-.dayStat {
-}
