--- web/src/components/Stat/DayStat/DayStat.tsx
@@ -1,7 +1,0 @@
-import "./DayStat.scss";
-
-const DayStat = (): JSX.Element => {
-  return <div className="dayStat">dayStat</div>;
-};
-
-export default DayStat;
