--- web/src/components/Stat/DayStat/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DayStat";
