--- web/src/components/Stat/MonthStat/MonthStat.scss
@@ -1,2 +1,0 @@
-.monthStat {
-}
