--- web/src/components/Stat/MonthStat/MonthStat.tsx
@@ -1,13 +1,0 @@
-import Calendar from "components/Common/Calendar";
-
-import "./MonthStat.scss";
-
-const MonthStat = (): JSX.Element => {
-  return (
-    <div className="monthStat">
-      <Calendar />
-    </div>
-  );
-};
-
-export default MonthStat;
