--- web/src/components/Stat/MonthStat/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./MonthStat";
