+++ web/src/components/Stat/StatItem/StatItem.scss
@@ -1,0 +1,12 @@
+.statItem {
+  &-left {
+    float: left;
+    width: 50%;
+    height: 100%;
+  }
+  &-right {
+    float: left;
+    width: 50%;
+    height: 100%;
+  }
+}
