+++ web/src/components/Stat/StatItem/WeekStat.tsx
@@ -1,0 +1,15 @@
+import Calendar from "components/Common/Calendar";
+
+import "./StatItem.scss";
+
+const WeekStat = (): JSX.Element => {
+  return (
+    <div className="statItem">
+      <div className="statItem-left">
+        <Calendar />
+      </div>
+      <div className="statItem-right"></div>
+    </div>
+  );
+};
+export default WeekStat;
