--- web/src/components/Stat/WeekStat/WeekStat.scss
@@ -1,2 +1,0 @@
-.weekStat {
-}
