--- web/src/components/Stat/WeekStat/WeekStat.tsx
@@ -1,7 +1,0 @@
-import "./WeekStat.scss";
-
-const WeekStat = (): JSX.Element => {
-  return <div className="weekStat">weekStat</div>;
-};
-
-export default WeekStat;
