--- web/src/components/Stat/WeekStat/index.ts
@@ -1,3 +1,0 @@
-import { formatWithOptions } from "util";
-
-export { default } from "./WeekStat";
